Use the external URL for link checking
This doesn't seem correct, as the callback should happen internally,
but the way the current routing is setup, there is a constraint in the
routing that means the request has to come in using the external
hostname.

At the moment, I'm assuming this is working in environments using the
same URLs for internal and external routing, and environments using
different URLs for internal and external routing, and this change
should keep the behaviour the same in both of those cases.

In the future however, it may be useful to switch this callback to use
the internal URLs, and change the routing to make that possible.

This uses the new external_url_for method in Plek, which handles
environments where the URLs for internal and external routing are
different.