Implement AssetManagerStorage::File#size
TODO: Add tests for new argument
TODO: Fetch real size from AM API when retrieve! is called?

The `size` method is defined on AssetManagerAndQuarantinedFileStorage
and on SanitizedFile and there are a number of places in the code
which expect it to be defined when we switch AttachmentUploader to use
AssetManagerStorage.

Unfortunately by the time the AssetManagerStorage::File is returned
from the `store!` method the file in the `asset_manager_tmp_dir` has
been deleted. When `store!` returns it also deletes the original file
passed in to the `store!` method, so we can't hold a reference to it
in `AssetManagerStorage::File`. Instead we calculate the file size in
`store!` and pass the value into the constructor of
`AssetManagerStorage::File`. We don't expect the file size to change
in the course of running the AssetManagerCreateWhitehallAssetWorker so
this should be fine.

We can't easily implement this in the `retrieve!` method unless we
call the Asset Manager API. For now we'll leave the value as `nil` in
this case and see if we need to do anything else.