Send access_limited data to Asset Manager
Three models that include `Attachable`: PolicyGroup, Response and
Edition.

`AttachmentData#attachable` arbitrarily takes the attachable object
belonging to the last attachment. In theory, all attachable objects
belonging to an AttachmentData will have the same parent Document but I
don't think there's anything in the model to enforce this.
  * TODO: I think I need to go up to the document and then find the
latest edition

* TODO: New commit - don't set access_limited on object creation; set it
when attachments are created and edition state changes.
  * I'll probably need to set an initial state to ensure that the asset
isn't visible to anyone until we've either set access_limited or
explicitly unset it
  * Could I have:
    * access_limited: true - not visible to anyone
    * access_limited: false - visible to all but draft state is
effective
    * access_limited: array of users - only visible to users in the
array

* Can't look up access_limited in `AssetManagerStorage#store!` as,
although the AttachmentData has been saved, its parent `Attachment`
hasn't, which means we can access the `Attachable` object.

* I think ideally we'd just store the (e.g.) AttachmentData class and ID
in the queue and have the Sidekiq job upload all versions of an
uploader-mounted file. We can't do this at the moment because of the
point we're hooking into CarrierWave.

* The `access_limited?` method added to `Response` and `PolicyGroup` is
to keep them consistent with the method in `Edition::LimitedAccess`.

* TODO: Add more unit tests around
AssetManagerCreateWhitehallAssetWorker#perform. To test the code I've
added to ensure it handles missing associations.